extra checks and forgot to cd to source directory

+++ phester/cronscripts/update_directory.sh
@@ -6,7 +6,7 @@
 # Copyright (C) 2002 Gerrit Riessen
 # This code is licensed under the GNU Public License.
 #
-# $Id: update_directory.sh,v 1.7 2002/05/29 10:37:49 riessen Exp $
+# $Id: update_directory.sh,v 1.8 2002/05/30 07:45:51 riessen Exp $
 #
 #
 # utility script for taking the latest phester output tar (generated
@@ -93,15 +93,19 @@ done
 echo "</ul></body></html>" >> index.html
 
 # remove the excessive tar files in the source directory
+cd $SRC_DIR
 if (( $TAR_FILES_KEEP > 0 )); 
 then
   remove_file_count=$(( `ls | wc -l` - $TAR_FILES_KEEP ))
-  echo "Removing last $remove_file_count files ..."
-  for n in `ls --sort=time | tail -$remove_file_count`; 
-  do
+  if (( $remove_file_count > 0 ));
+  then
+    echo "Removing last $remove_file_count files ..."
+    for n in `ls --sort=time | tail -$remove_file_count`; 
+    do
       echo "  removing $n ..."
       rm -f $n
-  done
+    done
+  fi
 fi
 
 # clean up after ourselves
